Rename file to make qt-creator compile successfully
